refactor: extract partition helpers
