+++ logs/log1756135976.md
@@ -1,0 +1,7 @@
+## PartitionPlacementActor refactor
+
+- Extracted rebalance publishing and handover logic into `PartitionPlacementRebalance` for clearer separation of cluster topology concerns.
+- Moved activation lifecycle management into `PartitionActivationLifecycle` to isolate spawning and termination handling.
+- Simplified `PartitionPlacementActor.ReceiveAsync` to delegate to the new helpers, keeping public API intact while improving readability and maintainability.
+
+Motivation: This separation of concerns makes the actor easier to reason about and test, reduces method complexity, and aligns with the project's preference for pure functions and modular design.
